+++ examples/configs/src/view-configs/eng.js
@@ -71,11 +71,35 @@ export const eng2019 = {
             obsType: 'cell',
           },
         },
+        {
+          fileType: 'figure.jpg',
+          url: 'https://upload.wikimedia.org/wikipedia/commons/thumb/1/11/A_dog_sitting_on_a_chair%2C_smiling.jpg/1024px-A_dog_sitting_on_a_chair%2C_smiling.jpg',
+          options: {
+            caption: 'Dog 1',
+          },
+          coordinationValues: {
+            figure: 'D1',
+          },
+        },
+        {
+          fileType: 'figure.jpg',
+          url: 'https://upload.wikimedia.org/wikipedia/commons/thumb/8/80/A_dog_-_36797913216.jpg/1920px-A_dog_-_36797913216.jpg',
+          options: {
+            caption: 'Dog 2',
+          },
+          coordinationValues: {
+            figure: 'D2',
+          },
+        },
       ],
     },
   ],
   initStrategy: 'auto',
   coordinationSpace: {
+    figure: {
+      A: 'D1',
+      B: 'D2',
+    },
     embeddingType: {
       TSNE: 't-SNE',
       UMAP: 'UMAP',
@@ -144,15 +168,9 @@ export const eng2019 = {
       h: 4,
     },
     {
-      component: 'scatterplot',
+      component: 'staticFigure',
       coordinationScopes: {
-        embeddingType: 'TSNE',
-        embeddingZoom: 'TSNE',
-        embeddingObsSetLabelsVisible: 'A',
-        embeddingObsSetLabelSize: 'A',
-        embeddingObsSetPolygonsVisible: 'A',
-        embeddingObsRadiusMode: 'A',
-        embeddingObsRadius: 'A',
+        figure: ['A', 'B'],
       },
       x: 0,
       y: 2,