+++ packages/constants-internal/src/constants.js
@@ -17,6 +17,7 @@ export const ViewType = {
   OBS_SET_SIZES: 'obsSetSizes',
   OBS_SET_FEATURE_VALUE_DISTRIBUTION: 'obsSetFeatureValueDistribution',
   FEATURE_VALUE_HISTOGRAM: 'featureValueHistogram',
+  STATIC_FIGURE: 'staticFigure',
 };
 
 export const DataType = {
@@ -30,6 +31,7 @@ export const DataType = {
   OBS_SEGMENTATIONS: 'obsSegmentations',
   NEIGHBORHOODS: 'neighborhoods',
   GENOMIC_PROFILES: 'genomic-profiles',
+  FIGURE: 'figure',
 };
 
 
@@ -99,6 +101,7 @@ export const FileType = {
   CLUSTERS_JSON: 'clusters.json',
   GENES_JSON: 'genes.json',
   ANNDATA_EXPRESSION_MATRIX_ZARR: 'anndata-expression-matrix.zarr',
+  FIGURE_JPG: 'figure.jpg',
 };
 
 /**