+++ packages/main/all/src/setup.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import { useState, useLayoutEffect } from 'react';
-import { VitS, registerPluginFileType } from '@vitessce/vit-s';
+import { VitS, registerPluginFileType, AbstractTwoStepLoader } from '@vitessce/vit-s';
 
 // Register view type plugins
 import { register as registerDescription } from '@vitessce/description';
@@ -12,6 +12,7 @@ import { register as registerHeatmap } from '@vitessce/heatmap';
 import { register as registerFeatureList } from '@vitessce/feature-list';
 import { register as registerLayerController } from '@vitessce/layer-controller';
 import { register as registerStatus } from '@vitessce/status';
+import { register as registerStaticFigure } from '@vitessce/static-figure';
 import { registerCellSetExpression, registerCellSetSizes, registerExpressionHistogram } from '@vitessce/statistical-plots';
 import { registerHiglass, registerGenomicProfiles } from '@vitessce/genomic-profiles';
 
@@ -82,6 +83,7 @@ function setup() {
   registerFeatureList();
   registerLayerController();
   registerStatus();
+  registerStaticFigure();
   // Statistical plots
   registerCellSetExpression();
   registerCellSetSizes();