+++ packages/vit-s/src/index.js
@@ -48,6 +48,7 @@ export {
   useFeatureSelection,
   useObsFeatureMatrixIndices,
   useMultiObsLabels,
+  useMultiFigures,
 
   useObsLocationsData,
   useObsSegmentationsData,