+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - Added `!important` statements to override Jupyter Notebook (classic) style conflicts.
 - Added `vite.config.js` and `pnpm run bundle` for bundling sub-packages as proper ESM.
 - Added test for a "consumer" package of the Vitessce sub-packages in `consumer/` directory, with a corresponding cypress test in `sites/html/cypress/e2e/html.spec.cy.js`.
+- Added `staticFigure` view to support showing static image files.
 
 ### Changed
 - Converted all `rem` units to `px` to fix R/Python widget CSS bugs caused by different root style conflicts.